+++ src/compiler/ast_compiler.c
@@ -1,0 +1,1 @@
+#include "ast_compiler.h"
