+++ src/compiler/ast_compiler.h
@@ -1,0 +1,4 @@
+#ifndef __AST_COMPILER_H__
+#define __AST_COMPILER_H__
+
+#endif
