+++ src/vm/core.c
@@ -8,7 +8,6 @@
 #include <sys/stat.h>
 #include "class.h"
 #include "common.h"
-#include "compiler.h"
 #include "gc.h"
 #include "header_obj.h"
 #include "meta_obj.h"
@@ -27,6 +26,10 @@
 #include <time.h>
 #include <unistd.h>
 #include "disassemble.h"
+#include "ast.h"
+
+#include "compiler.h"
+#include "one_pass_compiler.h"
 
 #include "core.script.inc"
 
@@ -121,11 +124,11 @@ static ObjThread* load_module(VM* vm, Value module_name, const char* module_code
         for (int i = 0; i < core->module_var_name.count; i++) {
             String name = core->module_var_name.datas[i];
             Value val = core->module_var_value.datas[i];
-            define_module_var(vm, module, name.str, name.len, val);
+            one_pass_define_module_var(vm, module, name.str, name.len, val);
         }
     }
 
-    ObjFn* fn = compile_module(vm, module, module_code);
+    ObjFn* fn = one_pass_compile_module(vm, module, module_code);
     push_tmp_root(vm, (ObjHeader*)fn);
 
 #ifdef DIS_ASM_CHUNK
@@ -179,7 +182,7 @@ int add_symbol(VM* vm, SymbolTable* table, const char* symbol, u32 len) {
 static Class* define_class(VM* vm, ObjModule* module, const char* name) {
     // 创建裸类并作为普通模块变量在模块中定义
     Class* class = class_new_raw(vm, name, 0);
-    define_module_var(vm, module, name, strlen(name), OBJ_TO_VALUE(class));
+    one_pass_define_module_var(vm, module, name, strlen(name), OBJ_TO_VALUE(class));
     return class;
 }
 
@@ -1904,12 +1907,12 @@ def_prim(Range_new_arg1) {
     ROBJ(objrange_new(vm, from, args[1].i32val, step));
 }
 
-inline static char* get_file_path(const char* module_name, int mode) {
+inline static char* get_file_path(const char* module_name, enum ImportRootType mode) {
     u32 root_dir_len = -1;
     const char* root_path = NULL;
     
     // mode 决定路径的前缀
-    if (mode == 0) {
+    if (mode == DEFAULT_ROOT) {
         // 默认根，先查找SPR_HOME，若没有设置，再按root_dir查找
         root_path = getenv("SPR_HOME");
         if (root_path != NULL) {
@@ -1918,10 +1921,10 @@ inline static char* get_file_path(const char* module_name, int mode) {
             root_dir_len = root_dir == NULL ? 0 : strlen(root_dir);
             root_path = root_dir;
         }
-    } else if (mode == 1) {
+    } else if (mode == STD_ROOT) {
         root_path = getenv("SPR_STD_LIB_PATH")?: getenv("HOME");
         root_dir_len = strlen(root_path);
-    } else if (mode == 2) {
+    } else if (mode == LIB_ROOT) {
         root_path = getenv("SPR_LIB_PATH")?: getenv("HOME");
         root_dir_len = strlen(root_path);
     }
@@ -1946,7 +1949,7 @@ inline static char* get_file_path(const char* module_name, int mode) {
     return path;
 }
 
-inline static char* read_module(const char* module_name, int mode) {
+inline static char* read_module(const char* module_name, enum ImportRootType mode) {
     char* module_path = get_file_path(module_name, mode);
     char* module_code = read_file(module_path);
     free(module_path);
@@ -1958,7 +1961,7 @@ inline static void print_str(const char* str) {
     fflush(stdout);
 }
 
-static Value import_module(VM* vm, Value module_name, int mode) {
+static Value import_module(VM* vm, Value module_name, enum ImportRootType mode) {
     // mode: 0. root_dir, 1. std, 2. lib
     if (!VALUE_IS_UNDEFINED(objmap_get(vm->all_module, module_name))) {
         return VT_TO_VALUE(VT_NULL);
@@ -2025,7 +2028,7 @@ def_prim(System_get_time) {
     RF64((double)now.tv_sec * 1000.0 + (double)now.tv_nsec / 1000000.0);
 }
 
-inline static bool import_module_core(VM* vm, Value* args, int mode) {
+inline static bool import_module_core(VM* vm, Value* args, enum ImportRootType mode) {
     if (!validate_str(vm, args[1])) {
         return false; // error
     }
@@ -2044,15 +2047,15 @@ inline static bool import_module_core(VM* vm, Value* args, int mode) {
 }
 
 def_prim(System_import_module) {
-    return import_module_core(vm, args, 0);
+    return import_module_core(vm, args, DEFAULT_ROOT);
 }
 
 def_prim(System_import_std_module) {
-    return import_module_core(vm, args, 1);
+    return import_module_core(vm, args, STD_ROOT);
 }
 
 def_prim(System_import_lib_module) {
-    return import_module_core(vm, args, 2);
+    return import_module_core(vm, args, LIB_ROOT);
 }
 
 def_prim(System_get_module_variable) {
